change component to onpush and break utils/models out into their own files for unit testing
