+++ projects/swimlane/ngx-ui/src/lib/components/calendar/calendar-month.type.ts
@@ -1,0 +1,3 @@
+import { CalenderDay } from './calendar-day.interface';
+
+export type CalendarMonth = CalenderDay[][];
