+++ projects/swimlane/ngx-ui/src/lib/components/calendar/calendar-view.type.ts
@@ -1,0 +1,1 @@
+export type CalendarView = 'year' | 'month' | 'date';
