+++ projects/swimlane/ngx-ui/src/lib/components/calendar/calendar.component.ts
@@ -1,7 +1,11 @@
-import { Component, Input, Output, EventEmitter, forwardRef, OnInit, ViewEncapsulation } from '@angular/core';
+import { Component, Input, Output, EventEmitter, forwardRef, OnInit, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import moment from 'moment-timezone';
-import { getMonth, CalenderDay, Month, getDecadeStartYear } from './calendar-utils';
+
+import { getMonth, getDecadeStartYear } from './utils';
+import { CalenderDay } from './calendar-day.interface';
+import { CalendarMonth } from './calendar-month.type';
+import { CalendarView } from './calendar-view.type';
 
 const CALENDAR_VALUE_ACCESSOR = {
   provide: NG_VALUE_ACCESSOR,
@@ -9,12 +13,12 @@ const CALENDAR_VALUE_ACCESSOR = {
   multi: true
 };
 
-type View = 'year' | 'month' | 'date';
-
 @Component({
   selector: 'ngx-calendar',
+  exportAs: 'ngxCalendar',
   providers: [CALENDAR_VALUE_ACCESSOR],
   encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
   styleUrls: ['./calendar.component.scss'],
   template: `
     <div class="ngx-calendar-wrap">
@@ -143,27 +147,27 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
   @Input() inputFormats: any[] = ['L', `LT`, 'L LT', moment.ISO_8601];
 
   @Input('minView')
-  set minView(val: View) {
+  set minView(val: CalendarView) {
     this._minView = val;
     this.validateView();
   }
 
-  get minView(): View {
+  get minView() {
     // default: 'date'
     return this._minView ? this._minView : 'date';
   }
 
   @Input('defaultView')
-  set defaultView(val: View) {
+  set defaultView(val: CalendarView) {
     this._defaultView = val;
     this.validateView();
   }
 
-  get defaultView(): View {
+  get defaultView() {
     return this._defaultView ? this._defaultView : this.minView;
   }
 
-  @Output() change: EventEmitter<any> = new EventEmitter();
+  @Output() change = new EventEmitter<Date>();
 
   get value(): Date {
     return this._value;
@@ -182,18 +186,26 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
   }
 
   activeDate: moment.Moment;
-  _value: Date;
-  weeks: Month;
-
-  currentView: View;
-  monthsList: string[] = moment.monthsShort();
-  // today's date
-  current: moment.Moment;
+  weeks: CalendarMonth;
+  currentView: CalendarView;
+  monthsList = moment.monthsShort();
   startYear: number;
-  _minView: View;
-  _defaultView: View;
 
-  changeViews(): void {
+  private _value: Date;
+  private _current: moment.Moment;
+  private _minView: CalendarView;
+  private _defaultView: CalendarView;
+
+  ngOnInit() {
+    this.activeDate = this.createMoment(this.value);
+    this.weeks = getMonth(this.activeDate);
+    this.monthsList = moment.monthsShort();
+    this._current = this.activeDate;
+    this.startYear = getDecadeStartYear(this._current.year());
+    this.validateView();
+  }
+
+  changeViews() {
     if (this.currentView === 'date') {
       this.currentView = 'month';
     } else if (this.currentView === 'month') {
@@ -201,29 +213,24 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
     } else if (this.currentView === 'year') {
       this.currentView = this.minView;
     }
+
     this.weeks = getMonth(this.activeDate);
   }
 
-  validateView(): void {
+  validateView() {
     const viewsList = ['date', 'month', 'year'];
+
     // date time picker precision validation
     if (!viewsList.includes(this.minView)) {
       this.minView = 'date';
     }
+
     // defaultView cannot be below minView
     if (viewsList.indexOf(this.minView) > viewsList.indexOf(this.defaultView)) {
       this.defaultView = this.minView;
     }
-    this.currentView = this.defaultView;
-  }
 
-  ngOnInit(): void {
-    this.activeDate = this.createMoment(this.value);
-    this.weeks = getMonth(this.activeDate);
-    this.monthsList = moment.monthsShort();
-    this.current = this.activeDate;
-    this.startYear = getDecadeStartYear(this.current.year());
-    this.validateView();
+    this.currentView = this.defaultView;
   }
 
   /**
@@ -246,7 +253,7 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
    */
   isCurrentMonth(month: string): boolean {
     const date = this.activeDate.clone().month(month);
-    return date.isSame(this.current, 'month') && date.isSame(this.current, 'year');
+    return date.isSame(this._current, 'month') && date.isSame(this._current, 'year');
   }
 
   /**
@@ -262,14 +269,15 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
    */
   isCurrentYear(year: number): boolean {
     const date = this.createMoment(this.value).year(year);
-    return date.isSame(this.current, 'year');
+    return date.isSame(this._current, 'year');
   }
 
   isDisabled(value: any, type: string): boolean {
     if (this.disabled) return true;
     if (!value) return false;
 
-    let date;
+    let date: moment.Moment;
+
     switch (type) {
       case 'day':
         date = value;
@@ -283,55 +291,60 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
       default:
         return false;
     }
+
     const isBeforeMin = this.minDate && date.isBefore(this.parseDate(this.minDate), type);
     const isAfterMax = this.maxDate && date.isAfter(this.parseDate(this.maxDate), type);
+
     return isBeforeMin || isAfterMax;
   }
 
-  onDayClick(day: CalenderDay): void {
+  onDayClick(day: CalenderDay) {
     this.activeDate = day.date.clone();
     this.value = this.activeDate.toDate();
+
     if (day.prevMonth || day.nextMonth) {
       this.weeks = getMonth(this.activeDate);
     }
   }
 
-  onMonthClick(month: string): void {
+  onMonthClick(month: string) {
     this.activeDate.month(month);
     this.value = this.activeDate.toDate();
+
     if (this.minView !== 'month') {
       this.currentView = 'date';
       this.weeks = getMonth(this.activeDate);
     }
   }
 
-  onYearClick(year: number): void {
+  onYearClick(year: number) {
     this.activeDate.year(year);
     this.value = this.activeDate.toDate();
+
     if (this.minView !== 'year') {
       this.currentView = 'month';
       this.weeks = getMonth(this.activeDate);
     }
   }
 
-  prevMonth(): void {
+  prevMonth() {
     const date = this.activeDate.clone();
     this.activeDate = date.subtract(1, 'month');
     this.weeks = getMonth(this.activeDate);
   }
 
-  nextMonth(): void {
+  nextMonth() {
     const date = this.activeDate.clone();
     this.activeDate = date.add(1, 'month');
     this.weeks = getMonth(this.activeDate);
   }
 
-  prevYear(): void {
+  prevYear() {
     const date = this.activeDate.clone();
     this.activeDate = date.subtract(1, 'year');
   }
 
-  nextYear(): void {
+  nextYear() {
     const date = this.activeDate.clone();
     this.activeDate = date.add(1, 'year');
   }
@@ -344,8 +357,9 @@ export class CalendarComponent implements OnInit, ControlValueAccessor {
     this.startYear = this.startYear + 20;
   }
 
-  writeValue(val: any): void {
+  writeValue(val: any) {
     const activeDate = this.createMoment(val);
+
     if (activeDate.isValid() && !activeDate.isSame(this.value, 'day')) {
       this.activeDate = activeDate;
       this.weeks = getMonth(this.activeDate);
