+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-days-for-month/index.ts
@@ -1,0 +1,1 @@
+export { getDaysForMonth } from './get-days-for-month.util';
