+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-decade-start-year/get-decade-start-year.util.ts
@@ -1,0 +1,7 @@
+/**
+ * Return decade start year
+ * @param year
+ */
+export function getDecadeStartYear(year: number) {
+  return year - ((year - 1) % 20);
+}
