+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-decade-start-year/index.ts
@@ -1,0 +1,1 @@
+export { getDecadeStartYear } from './get-decade-start-year.util';
