+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-month/index.ts
@@ -1,0 +1,1 @@
+export { getMonth } from './get-month.util';
