+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-number-range/get-number-range.util.ts
@@ -1,0 +1,16 @@
+/**
+ * Creates a range for the given start/finish
+ * @param start
+ * @param finish
+ * @return result of range
+ */
+export function getNumberRange(start: number, finish: number) {
+  const arr: number[] = [];
+  let i = start;
+
+  while (i < finish) {
+    arr.push(i++);
+  }
+
+  return arr;
+}
