+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-number-range/index.ts
@@ -1,0 +1,1 @@
+export { getNumberRange } from './get-number-range.util';
