+++ projects/swimlane/ngx-ui/src/lib/components/calendar/utils/get-weeks-for-days/index.ts
@@ -1,0 +1,1 @@
+export { getWeeksForDays } from './get-weeks-for-days.util';
